Moved out streams list to separate component - StreamsTable
