+++ packages/livepeer.com/www/components/TokenTable/index.tsx
@@ -4,10 +4,8 @@ import { Box, Button, Flex, Input } from "@theme-ui/components";
 import Modal from "../Modal";
 import { Table, TableRow, Checkbox } from "../Table";
 import CopyBox from "../CopyBox";
-import moment from 'moment'
 
 export default ({ userId, id }) => {
-  const [streams, setStreams] = useState([]);
   const [tokens, setTokens] = useState([]);
   const [tokenName, setTokenName] = useState("");
   const [createModal, setCreateModal] = useState(false);
@@ -16,25 +14,12 @@ export default ({ userId, id }) => {
   const [newToken, setNewToken] = useState(null);
   const [selectedToken, setSelectedToken] = useState(null);
   const [copyTime, setCopyTime] = useState(null);
-  const { getStreams, getApiTokens, createApiToken, deleteApiToken } = useApi();
+  const { getApiTokens, createApiToken, deleteApiToken } = useApi();
   useEffect(() => {
     getApiTokens(userId)
       .then(tokens => setTokens(tokens))
       .catch(err => console.error(err)); // todo: surface this
   }, [userId, newToken, deleteModal]);
-  useEffect(() => {
-    getStreams(userId)
-      .then(streams => setStreams(streams))
-      .catch(err => console.error(err)); // todo: surface this
-  }, [userId, newToken, deleteModal]);
-  useEffect(() => {
-    const interval = setInterval(() => {
-      getStreams(userId)
-        .then(streams => setStreams(streams))
-        .catch(err => console.error(err)); // todo: surface this
-    }, 10000);
-    return () => clearInterval(interval);
-  }, []);
   const close = () => {
     setCreateModal(false);
     setDeleteModal(false);
@@ -213,52 +198,6 @@ export default ({ userId, id }) => {
             </TableRow>
           );
         })}
-      </Table>
-      <p>
-        Streams:
-      </p>
-      <Table sx={{ gridTemplateColumns: "auto 1fr auto auto auto" }}>
-        <TableRow variant="header">
-          <Box>ID</Box>
-          <Box>Name</Box>
-          <Box>Details</Box>
-          <Box>Segments</Box>
-          <Box>Last Active</Box>
-        </TableRow>
-        {
-          streams.map(stream => {
-            const { id, name, lastSeen, sourceSegments, transcodedSegments,
-              presets, renditions } = stream;
-            let formattedLastSeen = <em>unseen</em>;
-            if (lastSeen) {
-              formattedLastSeen = (
-                <span>{
-                  moment
-                    .unix(lastSeen / 1000.0)
-                    .fromNow()}
-                </span>
-              );
-            }
-            let details = ''
-            if (presets?.length) {
-              details = `${presets}`
-            }
-            if (Object.keys(renditions||{}).length) {
-              if (details) {
-                details += '/'
-              }
-              details += `${renditions}`
-            }
-            return (
-              <TableRow key={id} >
-                <Box>{id}</Box>
-                <Box>{name}</Box>
-                <Box>{details}</Box>
-                <Box><span>{sourceSegments || 0}/{transcodedSegments || 0}</span></Box>
-                <Box>{formattedLastSeen}</Box>
-              </TableRow>
-            );
-          })}
       </Table>
     </Box>
   );
